# Add Comprehensive Advanced Optimizations Test Suite